requirements.txt
clipboard==0.0.4
pycryptodome==3.15.0

